[FE] Refactor : 메인 페이지 구조 변경
메인페이지에 dom관련 state 제거.
